+++ test/clojure_lsp/features/semantic_tokens_test.clj
@@ -43,7 +43,7 @@
 (deftest usage->absolute-token
   (is (= [6 3 3 1 -1]
          (#'semantic-tokens/element->absolute-token refered-usage-a
-                                                  :function))))
+                                                    :function))))
 
 (deftest absolute-token->relative-token
   (testing "without previous token"
@@ -64,7 +64,7 @@
                                                              (->token refered-usage-c :function))))))
 
 (deftest full-tokens
-  (testing "testing tokens order"
+  (testing "tokens order"
     (h/load-code-and-locs
       (code "(ns some.ns"
             "  (:require [foo.bar :as f :refer [baz]]))"
@@ -75,28 +75,72 @@
             ""
             "bla"
             "(comment)"))
-    (is (= [3 1 3 2 -1
-            5 1 7 2 -1]
+    (is (= [1 35 3 1 -1
+            2 1 3 2 -1
+            1 0 3 1 -1
+            1 0 1 0 -1
+            0 2 3 1 -1
+            2 0 3 1 -1
+            1 1 7 2 -1]
            (semantic-tokens/full-tokens "file:///a.clj"))))
-  (testing "testing macro refered tokens"
+  (testing "macro refered tokens"
     (h/load-code-and-locs
       (code "(ns some.ns (:require [clojure.test :refer [deftest]]))"
             "(deftest some-test 1)"))
     (is (= [0 44 7 2 -1
             1 1 7 2 -1]
            (semantic-tokens/full-tokens "file:///a.clj"))))
-  (testing "testing macro core tokens"
+  (testing "macro core tokens"
     (h/load-code-and-locs (code "(comment 1)"))
     (is (= [0 1 7 2 -1]
+           (semantic-tokens/full-tokens "file:///a.clj"))))
+  (testing "function declared tokens"
+    (h/load-code-and-locs (code "(def foo 1)"
+                                "foo"))
+    (is (= [0 1 3 2 -1
+            1 0 3 1 -1]
+           (semantic-tokens/full-tokens "file:///a.clj"))))
+  (testing "type alias for function tokens"
+    (h/load-code-and-locs (code "(ns some.ns (:require [foo.bar :as fb]))"
+                                "fb/some-foo-bar"))
+    (is (= [1 0 2 0 -1
+            0 3 12 1 -1]
+           (semantic-tokens/full-tokens "file:///a.clj"))))
+  (testing "java classes for function tokens"
+    (h/load-code-and-locs (code "(ns some.ns)"
+                                "^java.lang.String \"\""
+                                "String"
+                                "^String \"\""))
+    (is (= [1 1 16 1 -1
+            1 0 6 4 -1
+            1 1 6 4 -1]
+           (semantic-tokens/full-tokens "file:///a.clj"))))
+  (testing "keywords for keyword tokens"
+    (h/load-code-and-locs (code "(ns some.ns)"
+                                ":foo"
+                                ":some.ns/foo"
+                                "::foo"))
+    (is (= [1 0 4 3 -1
+            1 0 12 3 -1
+            1 0 5 3 -1]
+           (semantic-tokens/full-tokens "file:///a.clj"))))
+  (testing "locals destructuring for variable tokens"
+    (h/load-code-and-locs (code "(fn [{:keys [foo bar]}])"))
+    (is (= [0 1 2 2 -1
+            0 5 5 3 -1
+            0 0 5 3 -1
+            0 7 3 5 -1
+            0 4 3 5 -1]
            (semantic-tokens/full-tokens "file:///a.clj")))))
 
 (deftest range-tokens
-  (testing "testing tokens only for range"
+  (testing "tokens only for range"
     (h/load-code-and-locs
       (code "(ns some.ns (:require [foo.bar :refer [baz]]))"
             "(def bla baz)"
             "baz"
             "(comment)"
             "baz"))
-    (is (= [3 1 7 2 -1]
+    (is (= [2 0 3 1 -1
+            1 1 7 2 -1]
            (semantic-tokens/range-tokens "file:///a.clj" {:name-row 3 :name-col 0 :name-end-row 4 :name-end-col 0})))))